@@ -20,20 +19,20 @@ android {
             } else {
                 storeFile = file(System.getenv('TRAVIS_STORE_FILE'))
                 storePassword = System.getenv('TRAVIS_STORE_PASSWORD')
-                keyAlias System.getenv('TRAVIS_ALIAS_KEY')
+                keyAlias = System.getenv('TRAVIS_ALIAS_KEY')
                 keyPassword = System.getenv('TRAVIS_ALIAS_PASSWORD')
             }
         }
     }
 
-    compileSdkVersion 25
-    buildToolsVersion '25.0.3'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
     defaultConfig {
         applicationId "com.jugarte.gourmet"
         minSdkVersion 16
-        targetSdkVersion 25
-        versionCode 14
-        versionName "2.0.1"
+        targetSdkVersion 27
+        versionCode 15
+        versionName "2.0.2"
 
         vectorDrawables.useSupportLibrary = true
 
