@@ -7,12 +7,12 @@ repositories {
 }
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion '25.0.3'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 25
+        targetSdkVersion 27
         versionCode 1
         versionName "1.0"
     }
@@ -23,33 +23,38 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+
 }
 
 dependencies {
-    def supportLibVersion = '25.3.1'
-    compile "com.android.support:appcompat-v7:$supportLibVersion"
-    compile "com.android.support:support-v4:$supportLibVersion"
-    compile "com.android.support:design:$supportLibVersion"
+    def supportLibVersion = '28.0.0'
+    implementation "com.android.support:appcompat-v7:$supportLibVersion"
+    implementation "com.android.support:support-v4:$supportLibVersion"
+    implementation "com.android.support:design:$supportLibVersion"
 
     def firebaseVersion = '10.2.4'
-    compile "com.google.firebase:firebase-core:$firebaseVersion"
-    compile "com.google.firebase:firebase-crash:$firebaseVersion"
-    compile "com.google.firebase:firebase-database:$firebaseVersion"
-    compile "com.google.firebase:firebase-auth:$firebaseVersion"
+    implementation "com.google.firebase:firebase-core:$firebaseVersion"
+    implementation "com.google.firebase:firebase-crash:$firebaseVersion"
+    implementation "com.google.firebase:firebase-database:$firebaseVersion"
+    implementation "com.google.firebase:firebase-auth:$firebaseVersion"
 
-    compile 'com.squareup.okhttp3:okhttp:3.10.0'
-    compile "com.google.code.gson:gson:2.8.2"
+    implementation 'com.squareup.okhttp3:okhttp:3.10.0'
+    implementation "com.google.code.gson:gson:2.8.2"
 
     // RX
-    compile 'io.reactivex.rxjava2:rxandroid:2.0.2'
-    compile 'io.reactivex.rxjava2:rxjava:2.1.10'
-
+    implementation "io.reactivex.rxjava2:rxandroid:2.0.2"
+    implementation "io.reactivex.rxjava2:rxjava:2.1.5"
 
     def dagger2Version = '2.8'
-    compile "com.google.dagger:dagger:$dagger2Version"
+    implementation "com.google.dagger:dagger:$dagger2Version"
     annotationProcessor "com.google.dagger:dagger-compiler:$dagger2Version"
 
-    testCompile "junit:junit:4.12"
-    testCompile "org.mockito:mockito-core:2.7.21"
-    testCompile "org.json:json:20160810"
+    testImplementation "junit:junit:4.12"
+    testImplementation "org.mockito:mockito-core:2.7.21"
+    testImplementation "org.json:json:20160810"
 }
